+++ viikko6/my-app/src/views/Layout.jsx
@@ -1,25 +1,10 @@
-import {Link, Outlet} from 'react-router-dom';
+import {Outlet} from 'react-router-dom';
+import SiteNavigation from '../components/SiteNavigation';
 
 const Layout = () => (
   <div>
     <header>
-    <nav>
-    <ul>
-      <li>
-        <Link to="/">Home</Link>
-      </li>
-      <li>
-        <Link to="/profile">Profile</Link>
-      </li>
-      <li>
-        <Link to="/upload">Upload</Link>
-      </li>
-      <li>
-      <Link to="/login">Login</Link>
-      </li>
-      
-    </ul>
-  </nav>
+      <SiteNavigation />
     </header>
     <main>
       <Outlet />
