+++ viikko6/my-app/src/views/Profile.jsx
@@ -1,43 +1,15 @@
-import { useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import { useUserContext } from '../contexts/UserContext';
-import { useUser } from '../hooks/APiHooks';
-import UserData from '../components/UserData.jsx';
+import UserData from '../components/UserData';
 
-export const Profile = () => {
-  const { setUser } = useUserContext();
-  const { getUserByToken } = useUser();
+export const Profile = () => (
+  <div>
+    <h2 className="text-2xl font-bold">Tämä on minun profiilisivu</h2>
 
-  const getUser = async () => {
-    try {
-      const token = localStorage.getItem('token');
-      const userData = await getUserByToken(token);
-      setUser(userData.user);
-    } catch (error) {
-      console.log(error.message);
-    }
-  };
-
-  useEffect(() => {
-    getUser();
-  }, []);
-
-  return (
+    <p>
+      <Link to="/">Navigoi takaisin etusivulle</Link>
+    </p>
     <div>
-      <h2 className="text-2xl font-bold">Tämä on minun profiilisivu</h2>
-
-      <p>
-        <Link to="/">Navigoi takaisin etusivulle</Link>
-      </p>
-      <div>
-        <UserData />
-      </div>
+      <UserData />
     </div>
-  );
-};
-
-
-
-
-
-
+  </div>
+);
